+++ sbcc/ivr/input-handlers/olderEpisodesMenuHandler2.js
@@ -2,10 +2,13 @@ var notifyELK = require('../../../notifications/elk-notification/elkNotification
 var getAudioLink = require('../../utils/getAudioLink');
 var episodes = require('../../data/episodes');
 var invalidAttempts = require('../../utils/invalidAttempts');
+var addPlayedItem = require('../../utils/addPlayedItem');
 
 module.exports = function(input) {
     notifyELK();
     var lang = state.vars.lang;
+    call.vars.olderEpisodesMenuTwoCount = call.vars.olderEpisodesMenuTwoCount ? call.vars.olderEpisodesMenuTwoCount + 1 : 1;
+    var count = call.vars.olderEpisodesMenuTwoCount;
     var selectedEpisodeMenu = 'selected-episode-menu';
     var selectedEpisode;
 
@@ -15,6 +18,8 @@ module.exports = function(input) {
         selectedEpisode = episodes[episodes.length - 7].name;
         playAudio(getAudioLink(lang, selectedEpisode));
         state.vars.played = selectedEpisode;
+        addPlayedItem(selectedEpisode);
+        call.vars['olderEpisodesMenuTwoPlayed_' + count] = selectedEpisode;
         playAudio(getAudioLink(lang, selectedEpisodeMenu));
         promptKey('selectedTipOrEpisode2');
         break;
@@ -23,6 +28,8 @@ module.exports = function(input) {
         selectedEpisode = episodes[episodes.length - 8].name;
         playAudio(getAudioLink(lang, selectedEpisode));
         state.vars.played = selectedEpisode;
+        addPlayedItem(selectedEpisode);
+        call.vars['olderEpisodesMenuTwoPlayed_' + count] = selectedEpisode;
         playAudio(getAudioLink(lang, selectedEpisodeMenu));
         promptKey('selectedTipOrEpisode2');
         break;
@@ -31,6 +38,8 @@ module.exports = function(input) {
         selectedEpisode = episodes[episodes.length - 9].name;
         playAudio(getAudioLink(lang, selectedEpisode));
         state.vars.played = selectedEpisode;
+        addPlayedItem(selectedEpisode);
+        call.vars['olderEpisodesMenuTwoPlayed_' + count] = selectedEpisode;
         playAudio(getAudioLink(lang, selectedEpisodeMenu));
         promptKey('selectedTipOrEpisode2');
         break;
@@ -39,6 +48,8 @@ module.exports = function(input) {
         selectedEpisode = episodes[episodes.length - 10].name;
         playAudio(getAudioLink(lang, selectedEpisode));
         state.vars.played = selectedEpisode;
+        addPlayedItem(selectedEpisode);
+        call.vars['olderEpisodesMenuTwoPlayed_' + count] = selectedEpisode;
         playAudio(getAudioLink(lang, selectedEpisodeMenu));
         promptKey('selectedTipOrEpisode2');
         break;
@@ -47,6 +58,8 @@ module.exports = function(input) {
         selectedEpisode = episodes[episodes.length - 11].name;
         playAudio(getAudioLink(lang, selectedEpisode));
         state.vars.played = selectedEpisode;
+        addPlayedItem(selectedEpisode);
+        call.vars['olderEpisodesMenuTwoPlayed_' + count] = selectedEpisode;
         playAudio(getAudioLink(lang, selectedEpisodeMenu));
         promptKey('selectedTipOrEpisode2');
         break;
@@ -55,6 +68,8 @@ module.exports = function(input) {
         selectedEpisode = episodes[episodes.length - 12].name;
         playAudio(getAudioLink(lang, selectedEpisode));
         state.vars.played = selectedEpisode;
+        addPlayedItem(selectedEpisode);
+        call.vars['olderEpisodesMenuTwoPlayed_' + count] = selectedEpisode;
         playAudio(getAudioLink(lang, selectedEpisodeMenu));
         promptKey('selectedTipOrEpisode2');
         break;