+++ sbcc/ivr/input-handlers/topTipsMenuHandler1.js
@@ -2,10 +2,13 @@ var notifyELK = require('../../../notifications/elk-notification/elkNotification
 var getAudioLink = require('../../utils/getAudioLink');
 var topTips = require('../../data/tips');
 var invalidAttempts = require('../../utils/invalidAttempts');
+var addPlayedItem = require('../../utils/addPlayedItem');
 
 module.exports = function(input) {
     notifyELK();
     var lang = state.vars.lang;
+    call.vars.topTipsMenuOneCount = call.vars.topTipsMenuOneCount ? call.vars.topTipsMenuOneCount + 1 : 1;
+    var count = call.vars.topTipsMenuOneCount;
     var selectedTipMenu = 'selected-tip-menu';
     var selectedTip;
 
@@ -15,6 +18,8 @@ module.exports = function(input) {
         selectedTip = topTips[0].name;
         playAudio(getAudioLink(lang, selectedTip));
         state.vars.played = selectedTip;
+        addPlayedItem(selectedTip);
+        call.vars['topTipsMenuOnePlayed_' + count] = selectedTip;
         playAudio(getAudioLink(lang, selectedTipMenu));
         promptKey('selectedTipOrEpisode2');
         break;
@@ -23,6 +28,8 @@ module.exports = function(input) {
         selectedTip = topTips[1].name;
         playAudio(getAudioLink(lang, selectedTip));
         state.vars.played = selectedTip;
+        addPlayedItem(selectedTip);
+        call.vars['topTipsMenuOnePlayed_' + count] = selectedTip;
         playAudio(getAudioLink(lang, selectedTipMenu));
         promptKey('selectedTipOrEpisode2');
         break;
@@ -31,6 +38,8 @@ module.exports = function(input) {
         selectedTip = topTips[2].name;
         playAudio(getAudioLink(lang, selectedTip));
         state.vars.played = selectedTip;
+        addPlayedItem(selectedTip);
+        call.vars['topTipsMenuOnePlayed_' + count] = selectedTip;
         playAudio(getAudioLink(lang, selectedTipMenu));
         promptKey('selectedTipOrEpisode2');
         break;
@@ -39,6 +48,8 @@ module.exports = function(input) {
         selectedTip = topTips[3].name;
         playAudio(getAudioLink(lang, selectedTip));
         state.vars.played = selectedTip;
+        addPlayedItem(selectedTip);
+        call.vars['topTipsMenuOnePlayed_' + count] = selectedTip;
         playAudio(getAudioLink(lang, selectedTipMenu));
         promptKey('selectedTipOrEpisode2');
         break;